fix dimacode

+++ ImpostorHqR.Core/Web/Http/Server/Response/Fields/Fields.Content.Length.cs
@@ -8,9 +8,9 @@ namespace ImpostorHqR.Core.Web.Http.Server.Response.Fields
 
         public string Value => Convert.ToString(Size);
 
-        public int Size { get; }
+        public long Size { get; }
 
-        public FieldContentLength(int length) => this.Size = length;
+        public FieldContentLength(long length) => this.Size = length;
 
         public string Compile() => string.Concat(Code, Value);
     }
